add reducers to the controller definition
https://coveord.atlassian.net/browse/KIT-2866
